consolidate similar components into single file, move import schema modal into own component
