+++ stockpile-web/src/components/common/form/FormInputs.js
@@ -1,11 +1,31 @@
 import PropTypes from 'prop-types';
 import { getComponentId } from '../Util';
+
+/**
+ * Form text input field.
+ * @param {*} param0 
+ * @returns 
+ */
+const TextInput = ({id, label, placeholder, value, changeAction}) => {
+    return (
+        <div className="mb-3">
+            <label htmlFor={id} className="form-label">{label}</label>
+            <input type="text" className="form-control form-control-sm" id={id} value={value} placeholder={placeholder} onChange={changeAction} />
+        </div>
+    )
+}
+TextInput.propTypes = {
+    id: PropTypes.string.isRequired,
+    label: PropTypes.string,
+    changeAction: PropTypes.func
+}
+
 /**
  * Button component
  * @param {*} param0 
  * @returns 
  */
-const Button = ({id, form, type, className, label = 'Button', disabled, children, clickAction}) => {
+ const Button = ({id, form, type, className, label = 'Button', disabled, children, clickAction}) => {
     const buttonLabel = children ? children : label;
     const idValue = getComponentId(id);
 
@@ -15,7 +35,6 @@ const Button = ({id, form, type, className, label = 'Button', disabled, children
         </div>
     )
 }
-
 Button.propTypes = {
     children: PropTypes.node,
     clickAction: PropTypes.func,
@@ -24,10 +43,12 @@ Button.propTypes = {
     label: PropTypes.string,
     disabled: PropTypes.bool
 }
-
 Button.defaultProps = {
     type: 'button',
     disabled: false
 }
 
-export default Button
+export {
+    TextInput,
+    Button
+}
