--- stockpile-web/src/components/common/popup/PopupModalBody.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const PopupModalBody = ({children}) => {
-    return (
-        <div className="modal-body">
-            {children}
-        </div>
-    )
-}
-
-export default PopupModalBody
